NewTerrainANDFightSystem

+++ RPG/Assets/Scirpt/ItemSObject.cs
@@ -21,5 +21,5 @@ public class ItemSObject : ScriptableObject
     public float jumpCount;
     public float healthCount;
     public float manaCount;
-    public float damageCount;
+    public int damageCount;
 }
